restart_wazuh function has to be generic because is used in Mitre and FIM
