+++ test_wazuh/test_analysisd/conftest.py
@@ -9,25 +9,15 @@ import time
 
 import pytest
 
-from wazuh_testing.mitre import (LOG_FILE_PATH)
+from wazuh_testing.tools import restart_wazuh_service
 from wazuh_testing.tools import (FileMonitor, truncate_file,
                                  restart_wazuh_service)
 from wazuh_testing.tools import restart_wazuh_daemon
 
-@pytest.fixture(scope='module')
-def restart_wazuh(get_configuration, request):
-    # Reset ossec.log and start a new monitor
-    truncate_file(LOG_FILE_PATH)
-    file_monitor = FileMonitor(LOG_FILE_PATH)
-    setattr(request.module, 'wazuh_log_monitor', file_monitor)
-
-    # Restart Wazuh and wait for the command to end
-    restart_wazuh_service()
 
 @pytest.fixture(scope='module')
 def configure_local_rules(get_configuration, request):
     """Configure a custom rule in local_rules.xml for testing. Restart Wazuh is needed for applying the configuration."""
-    print(f"Test's path: {str(get_configuration)}")
 
     # save current configuration
     shutil.copy('/var/ossec/etc/rules/local_rules.xml', '/var/ossec/etc/rules/local_rules.xml.cpy')
